+++ CHANGELOG.md
@@ -16,6 +16,8 @@ The format is based on [Keep a Changelog][https://keepachangelog.com/en/1.0.0/] 
 
 ### Removed
 
+- Remove `Serialize` and `Deserialize` impls for `Option<T>`
+
 ### Changed
 
 ---